refactor(auth): Move and adapt AuthService to domain
