+++ apps/backend/users-service/src/auth/application/index.ts
@@ -1,0 +1,1 @@
+export * from './auth.service';
